Removing duplicate tests which are covered by the gherkin tests
 * Also renamed the remaining tests
 * Left most of the infrastructure in place in-case we want it again
 * We have lost some coverage of un-configured launches
 * The configured ones are covered better in gherkin though
