--- tests/functional/fixtures/lti_certification/v1.1/section_4.json
@@ -1,38 +1,0 @@
-{
-    "context_id": "con-182",
-    "context_label": "SI182",
-    "context_title": "Design of Personal Environments",
-    "context_type": "CourseSection",
-    "custom_context_memberships_url": "https://apps.imsglobal.org/lti/cert/tp/tp_membership.php/context/con-182/membership?b64=a3BwZGwzZGhvdGRsdHBkcDV2dWh1aDQzYzE%3D",
-    "custom_context_setting_url": "https://apps.imsglobal.org/lti/cert/tp/tp_settings.php/lis/CourseSection/con-182/bindings/ims/cert/custom?b64=a3BwZGwzZGhvdGRsdHBkcDV2dWh1aDQzYzE%3D",
-    "custom_link_setting_url": "https://apps.imsglobal.org/lti/cert/tp/tp_settings.php/links/rli-1234/custom?b64=a3BwZGwzZGhvdGRsdHBkcDV2dWh1aDQzYzE%3D",
-    "custom_system_setting_url": "https://apps.imsglobal.org/lti/cert/tp/tp_settings.php/ToolProxy/Hypothesis4dd96539c449ca5c3d57cc3d778d6bf3/custom?b64=a3BwZGwzZGhvdGRsdHBkcDV2dWh1aDQzYzE%3D",
-    "custom_tc_profile_url": "https://apps.imsglobal.org/lti/cert/tp/tp_tcprofile.php?b64=a3BwZGwzZGhvdGRsdHBkcDV2dWh1aDQzYzE%3D",
-    "launch_presentation_document_target": "iframe",
-    "launch_presentation_locale": "en_US",
-    "launch_presentation_return_url": "https://apps.imsglobal.org/lti/cert/tp/tp_return.php/basic-lti-launch-request",
-    "lis_course_section_sourcedid": "id-182",
-    "lis_person_contact_email_primary": "jane@school.edu",
-    "lis_person_name_family": "Lastname",
-    "lis_person_name_given": "Jane",
-    "lis_person_name_full": "Jane Q. Lastname",
-    "lis_person_sourcedid": "school.edu:jane",
-    "lti_message_type": "basic-lti-launch-request",
-    "lti_version": "LTI-1p0",
-    "oauth_callback": "about:blank",
-    "oauth_consumer_key": "Hypothesis4dd96539c449ca5c3d57cc3d778d6bf3",
-    "oauth_nonce": "085e110b26c4aa3a3bd6a80fa1dfea6f",
-    "oauth_signature": "BvpjpH48sKi3EFURGQ9sDmv+jzA=",
-    "oauth_signature_method": "HMAC-SHA1",
-    "oauth_timestamp": "1573757500",
-    "oauth_version": "1.0",
-    "resource_link_id": "rli-1234",
-    "resource_link_title": "Link 1234",
-    "roles": "Instructor",
-    "tool_consumer_info_product_family_code": "imsglc",
-    "tool_consumer_info_version": "1.1",
-    "tool_consumer_instance_description": "IMS Testing Description",
-    "tool_consumer_instance_guid": "IMS Testing",
-    "tool_consumer_instance_name": "IMS Testing Instance",
-    "user_id": "123456"
-}
